+++ src/main/java/guideme/internal/data/GuideMELanguageProvider.java
@@ -30,8 +30,7 @@ public final class GuideMELanguageProvider extends LanguageProvider {
                 "debug", "Debug",
                 "debug.tooltip", "Advanced Debugging Settings for Guide development",
                 "adaptiveScaling", "Adaptive UI Scaling",
-                "showDebugGuiOverlays", "Debug GUI Overlays"
-        );
+                "showDebugGuiOverlays", "Debug GUI Overlays");
         for (var entry : translations.entrySet()) {
             add("guideme.configuration." + entry.getKey(), entry.getValue());
         }